refactored code. Remove ImageResultFrame. Moved `createGUI` to `ImageComparison`.
