+++ build.gradle
@@ -23,11 +23,6 @@ jacocoTestReport {
         xml.enabled = true // coveralls plugin depends on xml format report
         html.enabled = true
     }
-    afterEvaluate {
-        classDirectories = files(classDirectories.files.collect {
-            fileTree(dir: it, exclude: 'ua.comparison.image/ImageResultFrame**')
-        })
-    }
 }
 
 defaultTasks << 'clean'
